Make a trait for content editing commands
