+++ src/Core/Content/Extensions/Tags/SetTags.php
@@ -2,18 +2,16 @@
 
 namespace Smolblog\Core\Content\Extensions\Tags;
 
-use Smolblog\Core\Content\Queries\UserCanEditContent;
+use Smolblog\Core\Content\EditContentCommandKit;
 use Smolblog\Framework\Messages\AuthorizableMessage;
 use Smolblog\Framework\Messages\Command;
-use Smolblog\Framework\Messages\Query;
-use Smolblog\Framework\Messages\StoppableMessageKit;
 use Smolblog\Framework\Objects\Identifier;
 
 /**
  * Set the tags for a piece of content.
  */
 class SetTags extends Command implements AuthorizableMessage {
-	use StoppableMessageKit;
+	use EditContentCommandKit;
 
 	/**
 	 * Create the command.
@@ -29,18 +27,5 @@ class SetTags extends Command implements AuthorizableMessage {
 		public readonly Identifier $contentId,
 		public readonly array $tags,
 	) {
-	}
-
-	/**
-	 * The user must be able to edit the content.
-	 *
-	 * @return Query
-	 */
-	public function getAuthorizationQuery(): Query {
-		return new UserCanEditContent(
-			userId: $this->userId,
-			siteId: $this->siteId,
-			contentId: $this->contentId,
-		);
 	}
 }
