+++ src/Core/Content/Types/Reblog/EditReblogComment.php
@@ -2,6 +2,7 @@
 
 namespace Smolblog\Core\Content\Types\Reblog;
 
+use Smolblog\Core\Content\EditContentCommandKit;
 use Smolblog\Core\Content\Queries\UserCanEditContent;
 use Smolblog\Framework\Messages\AuthorizableMessage;
 use Smolblog\Framework\Messages\Command;
@@ -13,7 +14,14 @@ use Smolblog\Framework\Objects\Identifier;
  * Change the comment text on a Reblog.
  */
 class EditReblogComment extends Command implements AuthorizableMessage {
-	use StoppableMessageKit;
+	use EditContentCommandKit;
+
+	/**
+	 * Mirror the reblogId into contentId so we can use the trait.
+	 *
+	 * @var Identifier
+	 */
+	private Identifier $contentId;
 
 	/**
 	 * Construct the command.
@@ -29,18 +37,6 @@ class EditReblogComment extends Command implements AuthorizableMessage {
 		public readonly Identifier $reblogId,
 		public readonly string $comment,
 	) {
-	}
-
-	/**
-	 * User must be able to edit this content.
-	 *
-	 * @return Query
-	 */
-	public function getAuthorizationQuery(): Query {
-		return new UserCanEditContent(
-			userId: $this->userId,
-			siteId: $this->siteId,
-			contentId: $this->reblogId,
-		);
+		$this->contentId = $this->reblogId;
 	}
 }
