+++ src/Core/Content/Types/Status/DeleteStatus.php
@@ -2,6 +2,7 @@
 
 namespace Smolblog\Core\Content\Types\Status;
 
+use Smolblog\Core\Content\EditContentCommandKit;
 use Smolblog\Core\Content\Queries\UserCanEditContent;
 use Smolblog\Framework\Messages\AuthorizableMessage;
 use Smolblog\Framework\Messages\Command;
@@ -13,7 +14,14 @@ use Smolblog\Framework\Objects\Identifier;
  * Delete a status and remove it from projections.
  */
 class DeleteStatus extends Command implements AuthorizableMessage {
-	use StoppableMessageKit;
+	use EditContentCommandKit;
+
+	/**
+	 * Map statusId to contentID so we can use the trait.
+	 *
+	 * @var Identifier
+	 */
+	private Identifier $contentId;
 
 	/**
 	 * Construct the command.
@@ -27,18 +35,6 @@ class DeleteStatus extends Command implements AuthorizableMessage {
 		public readonly Identifier $userId,
 		public readonly Identifier $statusId,
 	) {
-	}
-
-	/**
-	 * User must be able to edit this content.
-	 *
-	 * @return Query
-	 */
-	public function getAuthorizationQuery(): Query {
-		return new UserCanEditContent(
-			userId: $this->userId,
-			siteId: $this->siteId,
-			contentId: $this->statusId,
-		);
+		$this->contentId = $this->statusId;
 	}
 }
